+++ src/scalecast/__init__.py
@@ -1,4 +1,4 @@
-__version__ = '0.18.7'
+__version__ = '0.18.8'
 
 from .util import metrics
 import inspect